Add handling ipv6 packets with No Next Header set but with meaningless  payload
If the next header field is set to No Next Header (value 59) the ipv6 header
should not be followed by any header, not even a header of an upper layer
protocol. However there still can be data in the payload what can be deduced
from the payload length field. This data should be passed by routers.